products/requirements.txt
peewee
mysqlclient==2.1.1
grpcio-tools
grpcio